chore(common): CHECKOUT-6377 Add coupon, currency and discount model
