+++ packages/common-types/src/lib/coupon/coupon.ts
@@ -1,0 +1,7 @@
+export default interface Coupon {
+    id: string;
+    displayName: string;
+    code: string;
+    couponType: string;
+    discountedAmount: number;
+}
