+++ packages/common-types/src/lib/coupon/index.ts
@@ -1,0 +1,1 @@
+export { default as Coupon } from './coupon';
