+++ packages/common-types/src/lib/currency/index.ts
@@ -1,0 +1,1 @@
+export { default as Currency, CurrencyConfig } from './currency';
