+++ packages/common-types/src/lib/discount/discount.ts
@@ -1,0 +1,4 @@
+export default interface Discount {
+    id: string;
+    discountedAmount: number;
+}
