+++ packages/common-types/src/lib/discount/index.ts
@@ -1,0 +1,1 @@
+export { default as Discount } from './discount';
